+++ blocks/form/gdpr-confirmation.js
@@ -1,11 +1,4 @@
-// eslint-disable-next-line import/no-cycle
-import { configlIst } from './form.js';
-
-const accept = '/accept';
-const urlProd = configlIst[0];
-const urlBase = urlProd;
-
-/* get url parameters */
+// get url parameters
 const curUrl = window.location.href;
 const urlParams = new URLSearchParams(window.location.search);
 const paramdata = urlParams.get('data');
@@ -15,17 +8,9 @@ function hasQueryParams() {
   return curUrl.toString().includes('?');
 }
 
-if (!hasQueryParams(curUrl)) {
-  const btn = document.querySelector('.button');
-  btn.setAttribute('disabled', '');
-}
-
-/**
- * Send a query to the GDPR API and save the apiKey to localStorage for further calls
- * form fields */
-/* eslint-disable import/prefer-default-export */
-export async function acceptresult() {
-  const url = urlBase + accept;
+// Send a query parameters to the GDPR API Accept
+export async function acceptresult(accept, urlProd, secondConfig) {
+  const url = urlProd + accept;
   const options = {
     method: 'PUT',
     headers: { 'Content-type': 'application/json; charset=UTF-8' },
@@ -39,7 +24,7 @@ export async function acceptresult() {
     const response = await fetch(url, options);
     const resData = await response.text();
     if (response.ok) {
-      window.open(configlIst[1]);
+      window.open(secondConfig);
     } else {
       const { message } = JSON.parse(resData);
       throw message;
@@ -52,4 +37,42 @@ export async function acceptresult() {
       'color:white; background-color:#290002',
     );
   }
+}
+
+export function createConfirmForm(block) {
+  const confirmationForm = document.createElement('form');
+  const configlIst = [];
+  [...block.children].forEach((row) => {
+    configlIst.push(row.children[1].textContent);
+  });
+  block.textContent = '';
+  // creates a check box
+  const checkbox = document.createElement('input');
+  checkbox.setAttribute('type', 'checkbox');
+  checkbox.setAttribute('required', 'true');
+  checkbox.setAttribute('name', 'confirm');
+  checkbox.className = 'confirmation-checkbox';
+  const label = document.createElement('label');
+  label.setAttribute('for', 'confirm');
+  const [firstConfig, secondConfig, thirdConfig, fourthConfig] = configlIst;
+  label.innerHTML = fourthConfig;
+  label.className = 'confirm-label';
+  // creates a button
+  const btn = document.createElement('button');
+  btn.innerText = thirdConfig;
+  btn.className = 'button';
+  confirmationForm.method = 'POST';
+  confirmationForm.append(checkbox, label, btn);
+  block.append(confirmationForm);
+  // checking query parameters
+  if (!hasQueryParams(curUrl)) {
+    btn.setAttribute('disabled', '');
+  }
+  // API Variables
+  const accept = '/accept';
+  const urlProd = firstConfig;
+  confirmationForm.onsubmit = async (e) => {
+    e.preventDefault();
+    acceptresult(accept, urlProd, secondConfig);
+  };
 }
